# lumina/server/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lumina API",
    "description": "REST API for Lumina social media platform",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3002", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "bio": { "type": "string" },
          "avatar": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  },
  "paths": {
...
    "/api/account": {
      "get": {
        "tags": ["Account"],
        "summary": "Get account details",
        "description": "Retrieve current user account details",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Account details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "tags": ["Account"],
        "summary": "Update account",
        "description": "Update user account details",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "bio": { "type": "string" },
                  "avatar": { "type": "string" },
                  "username": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Account updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["Account"],
        "summary": "Delete account",
        "description": "Permanently delete user account",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": {
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Account deleted" },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/account/export": {
      "get": {
        "tags": ["Account"],
        "summary": "Export account data",
        "description": "Export all user data",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Exported data" },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/export": {
      "get": {
        "tags": ["Export"],
        "summary": "Export data",
        "description": "Export user data",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Exported data" },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}